remove depdendency on pkg/config everywhere
Remaining dependencies:

     github.com/DataDog/datadog-agent/pkg/config/remote
     github.com/DataDog/datadog-agent/pkg/config/remote/data
     github.com/DataDog/datadog-agent/pkg/obfuscate
     github.com/DataDog/datadog-agent/pkg/proto/pbgo
     github.com/DataDog/datadog-agent/pkg/tagger
     github.com/DataDog/datadog-agent/pkg/tagger/collectors
     github.com/DataDog/datadog-agent/pkg/util/grpc
     github.com/DataDog/datadog-agent/pkg/util/http
     github.com/DataDog/datadog-agent/pkg/util/profiling
