+++ cmd/trace-agent/config_darwin.go
@@ -3,13 +3,10 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build !windows && !darwin
-// +build !windows,!darwin
-
-package config
+package main
 
 // DefaultLogFilePath is where the agent will write logs if not overridden in the conf
-const DefaultLogFilePath = "/var/log/datadog/trace-agent.log"
+const DefaultLogFilePath = "/opt/datadog-agent/logs/trace-agent.log"
 
 // defaultDDAgentBin specifies the default path to the main agent binary.
 const defaultDDAgentBin = "/opt/datadog-agent/bin/agent/agent"
