+++ cmd/trace-agent/config_nix.go
@@ -3,10 +3,13 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package config
+//go:build !windows && !darwin
+// +build !windows,!darwin
+
+package main
 
 // DefaultLogFilePath is where the agent will write logs if not overridden in the conf
-const DefaultLogFilePath = "/opt/datadog-agent/logs/trace-agent.log"
+const DefaultLogFilePath = "/var/log/datadog/trace-agent.log"
 
 // defaultDDAgentBin specifies the default path to the main agent binary.
 const defaultDDAgentBin = "/opt/datadog-agent/bin/agent/agent"
