+++ cmd/trace-agent/flags/flags_windows.go
