+++ cmd/trace-agent/main_windows.go
@@ -16,9 +16,8 @@ import (
 	"path/filepath"
 	"time"
 
+	"github.com/DataDog/datadog-agent/cmd/trace-agent/flags"
 	"github.com/DataDog/datadog-agent/pkg/runtime"
-	"github.com/DataDog/datadog-agent/pkg/trace/agent"
-	"github.com/DataDog/datadog-agent/pkg/trace/flags"
 	"github.com/DataDog/datadog-agent/pkg/trace/watchdog"
 	_ "github.com/DataDog/datadog-agent/pkg/util/containers/providers/windows"
 
@@ -64,7 +63,7 @@ func (m *myservice) Execute(args []string, r <-chan svc.ChangeRequest, changes c
 		}
 	}()
 	elog.Info(0x40000003, ServiceName)
-	agent.Run(ctx)
+	Run(ctx)
 
 	changes <- svc.Status{State: svc.Stopped}
 	return
@@ -165,7 +164,7 @@ func main() {
 	}()
 
 	// Invoke the Agent
-	agent.Run(ctx)
+	Run(ctx)
 }
 
 func startService() error {
