+++ cmd/trace-agent/testdata/full.yaml
