+++ cmd/trace-agent/testdata/non-normalized_env_host_tag.yaml
