+++ cmd/trace-agent/testdata/ok_env_host_tag.yaml
