+++ cmd/trace-agent/testdata/ok_env_top_level.yaml
