+++ cmd/trace-agent/testdata/site_eu.yaml
