+++ cmd/trace-agent/testdata/stringcode.go.tmpl
