+++ cmd/trace-agent/testdata/undocumented.yaml
