+++ pkg/trace/api/pipeline_stats_test.go
@@ -142,7 +142,6 @@ func TestPipelineStatsProxyHandler(t *testing.T) {
 	})
 
 	t.Run("error", func(t *testing.T) {
-		defer mockConfig("site", "asd:\r\n")()
 		req, err := http.NewRequest("POST", "/some/path", nil)
 		if err != nil {
 			t.Fatal(err)
@@ -150,6 +149,7 @@ func TestPipelineStatsProxyHandler(t *testing.T) {
 		rec := httptest.NewRecorder()
 		conf := newTestReceiverConfig()
 		conf.Endpoints[0].Host = ""
+		conf.Site = "asd:\r\n"
 		r := newTestReceiverFromConfig(conf)
 		r.pipelineStatsProxyHandler().ServeHTTP(rec, req)
 		res := rec.Result()
