+++ pkg/trace/api/profiles.go
@@ -17,11 +17,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/DataDog/datadog-agent/pkg/config"
+	"github.com/DataDog/datadog-agent/pkg/trace/config"
 	"github.com/DataDog/datadog-agent/pkg/trace/info"
 	"github.com/DataDog/datadog-agent/pkg/trace/log"
 	"github.com/DataDog/datadog-agent/pkg/trace/metrics"
-	"github.com/DataDog/datadog-agent/pkg/util/fargate"
 )
 
 const (
@@ -36,16 +35,16 @@ const (
 // profilingEndpoints returns the profiling intake urls and their corresponding
 // api keys based on agent configuration. The main endpoint is always returned as
 // the first element in the slice.
-func profilingEndpoints(apiKey string) (urls []*url.URL, apiKeys []string, err error) {
+func profilingEndpoints(conf *config.AgentConfig) (urls []*url.URL, apiKeys []string, err error) {
 	main := profilingURLDefault
-	if v := config.Datadog.GetString("apm_config.profiling_dd_url"); v != "" {
+	if v := conf.ProfilingProxy.DDURL; v != "" {
 		main = v
 		if strings.HasSuffix(main, profilingV1EndpointSuffix) {
 			log.Warnf("The configured url %s for apm_config.profiling_dd_url is deprecated. "+
 				"The updated endpoint path is /api/v2/profile.", v)
 		}
-	} else if site := config.Datadog.GetString("site"); site != "" {
-		main = fmt.Sprintf(profilingURLTemplate, site)
+	} else if conf.Site != "" {
+		main = fmt.Sprintf(profilingURLTemplate, conf.Site)
 	}
 	u, err := url.Parse(main)
 	if err != nil {
@@ -53,10 +52,9 @@ func profilingEndpoints(apiKey string) (urls []*url.URL, apiKeys []string, err e
 		return nil, nil, fmt.Errorf("error parsing main profiling intake URL %s: %v", main, err)
 	}
 	urls = append(urls, u)
-	apiKeys = append(apiKeys, apiKey)
+	apiKeys = append(apiKeys, conf.APIKey())
 
-	if opt := "apm_config.profiling_additional_endpoints"; config.Datadog.IsSet(opt) {
-		extra := config.Datadog.GetStringMapStringSlice(opt)
+	if extra := conf.ProfilingProxy.AdditionalEndpoints; extra != nil {
 		for endpoint, keys := range extra {
 			u, err := url.Parse(endpoint)
 			if err != nil {
@@ -76,12 +74,12 @@ func profilingEndpoints(apiKey string) (urls []*url.URL, apiKeys []string, err e
 // If the main intake URL can not be computed because of config, the returned handler will always
 // return http.StatusInternalServerError along with a clarification.
 func (r *HTTPReceiver) profileProxyHandler() http.Handler {
-	targets, keys, err := profilingEndpoints(r.conf.APIKey())
+	targets, keys, err := profilingEndpoints(r.conf)
 	if err != nil {
 		return errorHandler(err)
 	}
 	tags := fmt.Sprintf("host:%s,default_env:%s,agent_version:%s", r.conf.Hostname, r.conf.DefaultEnv, info.Version)
-	if orch := r.conf.FargateOrchestrator; orch != fargate.Unknown {
+	if orch := r.conf.FargateOrchestrator; orch != config.OrchestratorUnknown {
 		tag := fmt.Sprintf("orchestrator:fargate_%s", strings.ToLower(string(orch)))
 		tags = tags + "," + tag
 	}
