+++ pkg/trace/config/config.go
@@ -6,27 +6,17 @@
 package config
 
 import (
-	"bytes"
-	"context"
 	"crypto/tls"
 	"errors"
-	"fmt"
 	"net"
 	"net/http"
 	"net/url"
-	"os"
-	"os/exec"
-	"strings"
+	"regexp"
 	"time"
 
-	"github.com/DataDog/datadog-agent/pkg/config"
-	coreconfig "github.com/DataDog/datadog-agent/pkg/config"
-	"github.com/DataDog/datadog-agent/pkg/proto/pbgo"
+	"github.com/DataDog/datadog-agent/pkg/obfuscate"
 	"github.com/DataDog/datadog-agent/pkg/trace/config/features"
 	"github.com/DataDog/datadog-agent/pkg/trace/log"
-	"github.com/DataDog/datadog-agent/pkg/util/fargate"
-	"github.com/DataDog/datadog-agent/pkg/util/grpc"
-	httputils "github.com/DataDog/datadog-agent/pkg/util/http"
 	"github.com/DataDog/datadog-agent/pkg/util/profiling"
 )
 
@@ -43,19 +33,243 @@ type Endpoint struct {
 	NoProxy bool
 }
 
+// TelemetryEndpointPrefix specifies the prefix of the telemetry endpoint URL.
+const TelemetryEndpointPrefix = "https://instrumentation-telemetry-intake."
+
+// OTLP holds the configuration for the OpenTelemetry receiver.
+type OTLP struct {
+	// BindHost specifies the host to bind the receiver to.
+	BindHost string `mapstructure:"-"`
+
+	// HTTPPort specifies the port to use for the plain HTTP receiver.
+	// If unset (or 0), the receiver will be off.
+	HTTPPort int `mapstructure:"http_port"`
+
+	// GRPCPort specifies the port to use for the plain HTTP receiver.
+	// If unset (or 0), the receiver will be off.
+	GRPCPort int `mapstructure:"grpc_port"`
+
+	// MaxRequestBytes specifies the maximum number of bytes that will be read
+	// from an incoming HTTP request.
+	MaxRequestBytes int64 `mapstructure:"-"`
+}
+
+// ObfuscationConfig holds the configuration for obfuscating sensitive data
+// for various span types.
+type ObfuscationConfig struct {
+	// ES holds the obfuscation configuration for ElasticSearch bodies.
+	ES JSONObfuscationConfig `mapstructure:"elasticsearch"`
+
+	// Mongo holds the obfuscation configuration for MongoDB queries.
+	Mongo JSONObfuscationConfig `mapstructure:"mongodb"`
+
+	// SQLExecPlan holds the obfuscation configuration for SQL Exec Plans. This is strictly for safety related obfuscation,
+	// not normalization. Normalization of exec plans is configured in SQLExecPlanNormalize.
+	SQLExecPlan JSONObfuscationConfig `mapstructure:"sql_exec_plan"`
+
+	// SQLExecPlanNormalize holds the normalization configuration for SQL Exec Plans.
+	SQLExecPlanNormalize JSONObfuscationConfig `mapstructure:"sql_exec_plan_normalize"`
+
+	// HTTP holds the obfuscation settings for HTTP URLs.
+	HTTP HTTPObfuscationConfig `mapstructure:"http"`
+
+	// RemoveStackTraces specifies whether stack traces should be removed.
+	// More specifically "error.stack" tag values will be cleared.
+	RemoveStackTraces bool `mapstructure:"remove_stack_traces"`
+
+	// Redis holds the configuration for obfuscating the "redis.raw_command" tag
+	// for spans of type "redis".
+	Redis Enablable `mapstructure:"redis"`
+
+	// Memcached holds the configuration for obfuscating the "memcached.command" tag
+	// for spans of type "memcached".
+	Memcached Enablable `mapstructure:"memcached"`
+
+	// CreditCards holds the configuration for obfuscating credit cards.
+	CreditCards CreditCardsConfig `mapstructure:"credit_cards"`
+}
+
+type AppSecConfig struct {
+	// Enabled reports whether AppSec is enabled.
+	Enabled bool
+	// MaxPayloadSize ...
+	MaxPayloadSize int64
+	// APIKey ...
+	APIKey string
+	// DDURL ...
+	DDURL string
+}
+
+// Export returns an obfuscate.Config matching o.
+func (o *ObfuscationConfig) Export() obfuscate.Config {
+	return obfuscate.Config{
+		SQL: obfuscate.SQLConfig{
+			TableNames:       features.Has("table_names"),
+			ReplaceDigits:    features.Has("quantize_sql_tables") || features.Has("replace_sql_digits"),
+			KeepSQLAlias:     features.Has("keep_sql_alias"),
+			DollarQuotedFunc: features.Has("dollar_quoted_func"),
+			Cache:            features.Has("sql_cache"),
+		},
+		ES: obfuscate.JSONConfig{
+			Enabled:            o.ES.Enabled,
+			KeepValues:         o.ES.KeepValues,
+			ObfuscateSQLValues: o.ES.ObfuscateSQLValues,
+		},
+		Mongo: obfuscate.JSONConfig{
+			Enabled:            o.Mongo.Enabled,
+			KeepValues:         o.Mongo.KeepValues,
+			ObfuscateSQLValues: o.Mongo.ObfuscateSQLValues,
+		},
+		SQLExecPlan: obfuscate.JSONConfig{
+			Enabled:            o.SQLExecPlan.Enabled,
+			KeepValues:         o.SQLExecPlan.KeepValues,
+			ObfuscateSQLValues: o.SQLExecPlan.ObfuscateSQLValues,
+		},
+		SQLExecPlanNormalize: obfuscate.JSONConfig{
+			Enabled:            o.SQLExecPlanNormalize.Enabled,
+			KeepValues:         o.SQLExecPlanNormalize.KeepValues,
+			ObfuscateSQLValues: o.SQLExecPlanNormalize.ObfuscateSQLValues,
+		},
+		HTTP: obfuscate.HTTPConfig{
+			RemoveQueryString: o.HTTP.RemoveQueryString,
+			RemovePathDigits:  o.HTTP.RemovePathDigits,
+		},
+		Logger: new(debugLogger),
+	}
+}
+
+type debugLogger struct{}
+
+func (debugLogger) Debugf(format string, params ...interface{}) {
+	log.Debugf(format, params...)
+}
+
+// CreditCardsConfig holds the configuration for credit card obfuscation in
+// (Meta) tags.
+type CreditCardsConfig struct {
+	// Enabled specifies whether this feature should be enabled.
+	Enabled bool `mapstructure:"enabled"`
+
+	// Luhn specifies whether Luhn checksum validation should be enabled.
+	// https://dev.to/shiraazm/goluhn-a-simple-library-for-generating-calculating-and-verifying-luhn-numbers-588j
+	// It reduces false positives, but increases the CPU time X3.
+	Luhn bool `mapstructure:"luhn"`
+}
+
+// HTTPObfuscationConfig holds the configuration settings for HTTP obfuscation.
+type HTTPObfuscationConfig struct {
+	// RemoveQueryStrings determines query strings to be removed from HTTP URLs.
+	RemoveQueryString bool `mapstructure:"remove_query_string" json:"remove_query_string"`
+
+	// RemovePathDigits determines digits in path segments to be obfuscated.
+	RemovePathDigits bool `mapstructure:"remove_paths_with_digits" json:"remove_path_digits"`
+}
+
+// Enablable can represent any option that has an "enabled" boolean sub-field.
+type Enablable struct {
+	Enabled bool `mapstructure:"enabled"`
+}
+
+// TelemetryConfig holds Instrumentation telemetry Endpoints information
+type TelemetryConfig struct {
+	Enabled   bool `mapstructure:"enabled"`
+	Endpoints []*Endpoint
+}
+
+// JSONObfuscationConfig holds the obfuscation configuration for sensitive
+// data found in JSON objects.
+type JSONObfuscationConfig struct {
+	// Enabled will specify whether obfuscation should be enabled.
+	Enabled bool `mapstructure:"enabled"`
+
+	// KeepValues will specify a set of keys for which their values will
+	// not be obfuscated.
+	KeepValues []string `mapstructure:"keep_values"`
+
+	// ObfuscateSQLValues will specify a set of keys for which their values
+	// will be passed through SQL obfuscation
+	ObfuscateSQLValues []string `mapstructure:"obfuscate_sql_values"`
+}
+
+// ReplaceRule specifies a replace rule.
+type ReplaceRule struct {
+	// Name specifies the name of the tag that the replace rule addresses. However,
+	// some exceptions apply such as:
+	// • "resource.name" will target the resource
+	// • "*" will target all tags and the resource
+	Name string `mapstructure:"name"`
+
+	// Pattern specifies the regexp pattern to be used when replacing. It must compile.
+	Pattern string `mapstructure:"pattern"`
+
+	// Re holds the compiled Pattern and is only used internally.
+	Re *regexp.Regexp `mapstructure:"-"`
+
+	// Repl specifies the replacement string to be used when Pattern matches.
+	Repl string `mapstructure:"repl"`
+}
+
+// WriterConfig specifies configuration for an API writer.
+type WriterConfig struct {
+	// ConnectionLimit specifies the maximum number of concurrent outgoing
+	// connections allowed for the sender.
+	ConnectionLimit int `mapstructure:"connection_limit"`
+
+	// QueueSize specifies the maximum number or payloads allowed to be queued
+	// in the sender.
+	QueueSize int `mapstructure:"queue_size"`
+
+	// FlushPeriodSeconds specifies the frequency at which the writer's buffer
+	// will be flushed to the sender, in seconds. Fractions are permitted.
+	FlushPeriodSeconds float64 `mapstructure:"flush_period_seconds"`
+}
+
+// FargateOrchestratorName is a Fargate orchestrator name.
+type FargateOrchestratorName string
+
+const (
+	// ECS represents AWS ECS
+	OrchestratorECS FargateOrchestratorName = "ECS"
+	// EKS represents AWS EKS
+	OrchestratorEKS FargateOrchestratorName = "EKS"
+	// Unknown is used when we cannot retrieve the orchestrator
+	OrchestratorUnknown FargateOrchestratorName = "Unknown"
+)
+
+// ProfilingProxyConfig ...
+type ProfilingProxyConfig struct {
+	// DDURL ...
+	DDURL string
+	// AdditionalEndpoints ...
+	AdditionalEndpoints map[string][]string
+}
+
+// DebuggerProxyConfig ...
+type DebuggerProxyConfig struct {
+	// DDURL ...
+	DDURL string
+	// APIKey ...
+	APIKey string
+}
+
 // AgentConfig handles the interpretation of the configuration (with default
 // behaviors) in one place. It is also a simple structure to share across all
 // the Agent components, with 100% safe and reliable values.
 // It is exposed with expvar, so make sure to exclude any sensible field
 // from JSON encoding. Use New() to create an instance.
 type AgentConfig struct {
-	Enabled             bool
-	FargateOrchestrator fargate.OrchestratorName
+	Enabled      bool
+	AgentVersion string
+	Site         string // the intake site to use (e.g. "datadoghq.com")
+
+	// FargateOrchestrator specifies the name of the Fargate orchestrator. e.g. "ECS", "EKS", "Unknown"
+	FargateOrchestrator FargateOrchestratorName
 
 	// Global
 	Hostname   string
 	DefaultEnv string // the traces will default to this environment
 	ConfigPath string // the source of this config, if any
+	AuthToken  string // from the auth token file
 
 	// Endpoints specifies the set of hosts and API keys where traces and stats
 	// will be uploaded to. The first endpoint is the main configuration endpoint;
@@ -83,6 +297,12 @@ type AgentConfig struct {
 	ReceiverTimeout int
 	MaxRequestBytes int64 // specifies the maximum allowed request size for incoming trace payloads
 
+	WindowsPipeName        string
+	PipeBufferSize         int
+	PipeSecurityDescriptor string
+
+	GUIPort string // the port of the Datadog Agent GUI (for control access)
+
 	// Writers
 	SynchronousFlushing     bool // Mode where traces are only submitted when FlushAsync is called, used for Serverless Extension
 	StatsWriter             *WriterConfig
@@ -90,8 +310,10 @@ type AgentConfig struct {
 	ConnectionResetInterval time.Duration // frequency at which outgoing connections are reset. 0 means no reset is performed
 
 	// internal telemetry
-	StatsdHost string
-	StatsdPort int
+	StatsdHost     string
+	StatsdPort     int
+	StatsdPipeName string // for Windows Pipes
+	StatsdSocket   string // for UDS Sockets
 
 	// logging
 	LogLevel      string
@@ -139,8 +361,25 @@ type AgentConfig struct {
 	// Profiling settings, or nil if profiling is disabled
 	ProfilingSettings *profiling.Settings
 
+	// ProfilingProxy specifies settings for the profiling proxy.
+	ProfilingProxy ProfilingProxyConfig
+
 	// Telemetry settings
 	TelemetryConfig *TelemetryConfig
+
+	// AppSec contains AppSec configuration.
+	AppSec AppSecConfig
+
+	// DebuggerProxy contains the settings for the Live Debugger proxy.
+	DebuggerProxy DebuggerProxyConfig
+
+	// Proxy specifies a function to return a proxy for a given Request.
+	// See (net/http.Transport).Proxy for more details.
+	Proxy func(*http.Request) (*url.URL, error) `json:"-"`
+
+	// MaxCatalogEntries specifies the maximum number of services to be added to the priority sampler's
+	// catalog. If not set (0) it will default to 5000.
+	MaxCatalogEntries int
 }
 
 // Tag represents a key/value pair.
@@ -150,17 +389,13 @@ type Tag struct {
 
 // New returns a configuration with the default values.
 func New() *AgentConfig {
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	orch := fargate.GetOrchestrator(ctx)
-	cancel()
-	if err := ctx.Err(); err != nil && err != context.Canceled {
-		log.Errorf("Failed to get Fargate orchestrator. This may cause issues if you are in a Fargate instance: %v", err)
-	}
 	return &AgentConfig{
 		Enabled:             true,
-		FargateOrchestrator: orch,
 		DefaultEnv:          "none",
 		Endpoints:           []*Endpoint{{Host: "https://trace.agent.datadoghq.com"}},
+		FargateOrchestrator: OrchestratorUnknown,
+		Site:                "datadoghq.com",
+		MaxCatalogEntries:   5000,
 
 		BucketInterval: time.Duration(10) * time.Second,
 
@@ -170,9 +405,12 @@ func New() *AgentConfig {
 		MaxEPS:          200,
 		MaxRemoteTPS:    100,
 
-		ReceiverHost:    "localhost",
-		ReceiverPort:    8126,
-		MaxRequestBytes: 50 * 1024 * 1024, // 50MB
+		ReceiverHost:           "localhost",
+		ReceiverPort:           8126,
+		MaxRequestBytes:        50 * 1024 * 1024, // 50MB
+		PipeBufferSize:         1_000_000,
+		PipeSecurityDescriptor: "D:AI(A;;GA;;;WD)",
+		GUIPort:                "5002",
 
 		StatsWriter:             new(WriterConfig),
 		TraceWriter:             new(WriterConfig),
@@ -182,7 +420,6 @@ func New() *AgentConfig {
 		StatsdPort: 8125,
 
 		LogLevel:      "INFO",
-		LogFilePath:   DefaultLogFilePath,
 		LogThrottling: true,
 
 		MaxMemory:        5e8, // 500 Mb, should rarely go above 50 Mb
@@ -196,10 +433,14 @@ func New() *AgentConfig {
 
 		GlobalTags: make(map[string]string),
 
-		DDAgentBin:   defaultDDAgentBin,
+		Proxy:        http.ProxyFromEnvironment,
 		OTLPReceiver: &OTLP{},
 		TelemetryConfig: &TelemetryConfig{
-			Endpoints: []*Endpoint{{Host: telemetryEndpointPrefix + coreconfig.DefaultSite}},
+			Endpoints: []*Endpoint{{Host: TelemetryEndpointPrefix + "datadoghq.com"}},
+		},
+		AppSec: AppSecConfig{
+			Enabled:        true,
+			MaxPayloadSize: 5 * 1024 * 1024,
 		},
 	}
 }
@@ -210,83 +451,6 @@ func (c *AgentConfig) APIKey() string {
 		return ""
 	}
 	return c.Endpoints[0].APIKey
-}
-
-// Validate validates if the current configuration is good for the agent to start with.
-func (c *AgentConfig) validate() error {
-	if len(c.Endpoints) == 0 || c.Endpoints[0].APIKey == "" {
-		return ErrMissingAPIKey
-	}
-	if c.DDAgentBin == "" {
-		return errors.New("agent binary path not set")
-	}
-	if c.Hostname == "" {
-		// no user-set hostname, try to acquire
-		if err := c.acquireHostname(); err != nil {
-			log.Debugf("Could not get hostname via gRPC: %v. Falling back to other methods.", err)
-			if err := c.acquireHostnameFallback(); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// fallbackHostnameFunc specifies the function to use for obtaining the hostname
-// when it can not be obtained by any other means. It is replaced in tests.
-var fallbackHostnameFunc = os.Hostname
-
-// acquireHostname attempts to acquire a hostname for the trace-agent by connecting to the core agent's
-// gRPC endpoints. If it fails, it will return an error.
-func (c *AgentConfig) acquireHostname() error {
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-	client, err := grpc.GetDDAgentClient(ctx)
-	if err != nil {
-		return err
-	}
-	reply, err := client.GetHostname(ctx, &pbgo.HostnameRequest{})
-	if err != nil {
-		return err
-	}
-	if features.Has("disable_empty_hostname") && reply.Hostname == "" {
-		log.Debugf("Acquired empty hostname from gRPC but it's disallowed.")
-		return errors.New("empty hostname disallowed")
-	}
-	c.Hostname = reply.Hostname
-	log.Debugf("Acquired hostname from gRPC: %s", c.Hostname)
-	return nil
-}
-
-// acquireHostnameFallback attempts to acquire a hostname for this configuration. It
-// tries to shell out to the infrastructure agent for this, if DD_AGENT_BIN is
-// set, otherwise falling back to os.Hostname.
-func (c *AgentConfig) acquireHostnameFallback() error {
-	var out bytes.Buffer
-	cmd := exec.Command(c.DDAgentBin, "hostname")
-	cmd.Env = append(os.Environ(), cmd.Env...) // needed for Windows
-	cmd.Stdout = &out
-	err := cmd.Run()
-	c.Hostname = strings.TrimSpace(out.String())
-	if emptyDisallowed := features.Has("disable_empty_hostname") && c.Hostname == ""; err != nil || emptyDisallowed {
-		if emptyDisallowed {
-			log.Debugf("Core agent returned empty hostname but is disallowed by disable_empty_hostname feature flag. Falling back to os.Hostname.")
-		}
-		// There was either an error retrieving the hostname from the core agent, or
-		// it was empty and its disallowed by the disable_empty_hostname feature flag.
-		host, err2 := fallbackHostnameFunc()
-		if err2 != nil {
-			return fmt.Errorf("couldn't get hostname from agent (%q), nor from OS (%q). Try specifying it by means of config or the DD_HOSTNAME env var", err, err2)
-		}
-		if emptyDisallowed && host == "" {
-			return errors.New("empty hostname disallowed")
-		}
-		c.Hostname = host
-		log.Debugf("Acquired hostname from OS: %q. Core agent was unreachable at %q: %v.", c.Hostname, c.DDAgentBin, err)
-		return nil
-	}
-	log.Debugf("Acquired hostname from core agent (%s): %q.", c.DDAgentBin, c.Hostname)
-	return nil
 }
 
 // NewHTTPClient returns a new http.Client to be used for outgoing connections to the
@@ -304,7 +468,7 @@ func (c *AgentConfig) NewHTTPTransport() *http.Transport {
 	transport := &http.Transport{
 		TLSClientConfig: &tls.Config{InsecureSkipVerify: c.SkipSSLValidation},
 		// below field values are from http.DefaultTransport (go1.12)
-		Proxy: http.ProxyFromEnvironment,
+		Proxy: c.Proxy,
 		DialContext: (&net.Dialer{
 			Timeout:   30 * time.Second,
 			KeepAlive: 30 * time.Second,
@@ -315,34 +479,5 @@ func (c *AgentConfig) NewHTTPTransport() *http.Transport {
 		TLSHandshakeTimeout:   10 * time.Second,
 		ExpectContinueTimeout: 1 * time.Second,
 	}
-	if p := coreconfig.GetProxies(); p != nil {
-		transport.Proxy = httputils.GetProxyTransportFunc(p)
-	}
 	return transport
-}
-
-// Load returns a new configuration based on the given path. The path must not necessarily exist
-// and a valid configuration can be returned based on defaults and environment variables. If a
-// valid configuration can not be obtained, an error is returned.
-func Load(path string) (*AgentConfig, error) {
-	cfg, err := prepareConfig(path)
-	if err != nil {
-		if !os.IsNotExist(err) {
-			return nil, err
-		}
-	} else {
-		log.Infof("Loaded configuration: %s", cfg.ConfigPath)
-	}
-	cfg.applyDatadogConfig()
-	return cfg, cfg.validate()
-}
-
-func prepareConfig(path string) (*AgentConfig, error) {
-	cfg := New()
-	config.Datadog.SetConfigFile(path)
-	if _, err := config.Load(); err != nil {
-		return cfg, err
-	}
-	cfg.ConfigPath = path
-	return cfg, nil
 }
